Rename src/osp/link/machines.* to src/osp/link/link.* and move src/adera/machine/links.* to src/adera
