+++ src/adera/links.h
@@ -26,7 +26,7 @@
 
 #include <osp/core/math_types.h>
 
-#include <osp/link/machines.h>
+#include <osp/link/link.h>
 #include <osp/link/signal.h>
 
 namespace adera
