+++ src/osp/link/signal.h
@@ -24,7 +24,7 @@
  */
 #pragma once
 
-#include "machines.h"
+#include "link.h"
 
 namespace osp::link
 {
